refactor: divide paymaster modal into several pieces
