+++ frontend/components/paymaster/common/index.ts
@@ -1,0 +1,3 @@
+import Line from "./Line";
+
+export { Line };
